requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
drf-excel==2.1.0
et-xmlfile==1.1.0
model-bakery==1.4.0
openpyxl==3.0.9
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
unicodecsv==0.14.1
